refactor: ♻️ #31 BetSceneを共通化する
